% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_palette.R
\name{iucn_palette}
\alias{iucn_palette}
\title{An IUCN palette generator}
\usage{
iucn_palette(category = "All", exclude = NA)
}
\arguments{
\item{category}{Category code to return. Defaults to "All" to return all categories. Choices are: \code{All}, \code{CO}, \code{CR}, \code{EN}, \code{VU}, \code{NT}, \code{LC},  \code{DD}, \code{NE}.}

\item{exclude}{Category code to exclude from palette. Defaults to NA. Choices are same as \code{category} excepot for \code{All}.}
}
\value{
A vector of colours.
}
\description{
This function returns official IUCN colour codes
}
\examples{
iucn_palette()
iucn_palette("CR")
iucn_palette(c("CR", "EN", "VU"))
iucn_palette(category="All", exclude="NE")
iucn_palette(category="All", exclude=c("NE", "CO"))
}
\keyword{colors}
\keyword{colours}
