% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_palettes.R
\docType{data}
\name{iucn_palettes}
\alias{iucn_palettes}
\title{An IUCN palette generator}
\format{An object of class \code{list} of length 9.}
\usage{
iucn_palettes
}
\description{
Complete named list of IUCN colour codes individually and full palette.
}
\details{
Use \code{\link{iucn_palette}} to construct palettes of desired IUCN categories
}
\keyword{datasets}
